Adição de annotations para database nas classes que serão persistidas + refatorações de nomes de códigos (padronização)

+++ disque-saude-backend-master/src/main/java/com/ufcg/si1/service/EspecialidadeService.java
@@ -23,7 +23,7 @@ public interface EspecialidadeService {
     public void insere(Especialidade esp)throws Rep,
             ObjetoJaExistenteException;
 
-    public boolean existe(int codigo);
+    public boolean existe(long codigo);
 
     public Especialidade findById(long id);
 }
